+++ CHANGELOG.md
@@ -11,3 +11,4 @@ This file tracks weekly progress, features completed, and milestones.
 - THe hooks are added to automate my CHANGELOG.md file
 - Streamlit: Sample webpage design is done
 - Basic Model Creation: We have created a base model with resnet, its the base
+- Hu Moments: Hu moments are created and a model with hu moments is being used to predict it. models_with_hu